day_5/requirements.txt
chainlit==2.9.5
aiohttp==3.13.3
python-dotenv==1.0.1